streamlit-hello-world-app/requirements.txt
streamlit~=1.38.0
databricks-sql-connector
google-generativeai
pandas